week2/project/restaurant_app/requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
psycopg2-binary==2.9.7
python-dotenv==1.1.1
